requirements.txt
feedparser
duckduckgo-search
newspaper3k
openai
lxml_html_clean
google.genai
ddgs
requests
